# backend/docs/swagger.yaml
basePath: /
definitions:
  activity.Activity:
    properties:
      activity_image:
        type: string
      creator_id:
        type: string
      date:
        type: string
      description:
        type: string
      id:
        type: string
      title:
        type: string
    type: object
...
  group.Group:
    properties:
      creator_id:
        type: string
      description:
        type: string
      end_date:
        type: string
      group_image:
        type: string
      id:
        type: string
      name:
        type: string
      start_date:
        type: string
    type: object
...
  responses.CreateInviteRequest:
    properties:
      creator_id:
        example: user123
        type: string
      expires_at:
        example: "2025-12-31T23:59:59Z"
        type: string
    type: object
...
  responses.ErrorResponse:
    properties:
      error:
        example: Invalid request
        type: string
    type: object
  responses.GroupCreateRequest:
    properties:
      description:
        example: A group for studying algorithms
        type: string
      end_date:
        example: "2025-12-31"
        type: string
      group_image:
        example: https://example.com/image.jpg
        type: string
      name:
        example: Study Group
        type: string
    type: object
  responses.GroupDeleteRequest:
    properties:
      creator_id:
        example: user123
        type: string
    type: object
...
  responses.GroupUpdateRequest:
    properties:
      description:
        example: Updated description
        type: string
      end_date:
        example: "2025-12-31"
        type: string
      group_image:
        example: https://example.com/image.jpg
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API for CODECK - COmpetição DE Código do Kim. A platform for competitive
    programmers.
  license:
    name: MIT
    url: https://github.com/your-username/codeck/blob/main/LICENSE
  termsOfService: http://swagger.io/terms/
  title: CODECK API
  version: "1.0"
paths:
...
  /groups:
    post:
      consumes:
      - application/json
      description: Create a new group with name, end date, and optional image/description
      parameters:
      - description: Group creation data
        in: body
        name: group
        required: true
        schema:
          $ref: '#/definitions/responses.GroupCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/group.Group'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      summary: Create a new group
      tags:
      - groups
  /groups/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a group (only creator can delete)
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      - description: Delete request with creator_id
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/responses.GroupDeleteRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      summary: Delete a group
      tags:
      - groups
    get:
      consumes:
      - application/json
      description: Get group information (members only)
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      - description: Requester User ID
        in: query
        name: requester_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/group.Group'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      summary: Get group details
      tags:
      - groups
    put:
      consumes:
      - application/json
      description: Update group information (name cannot be updated)
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      - description: Group update data
        in: body
        name: group
        required: true
        schema:
          $ref: '#/definitions/responses.GroupUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/group.Group'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      summary: Update an existing group
      tags:
      - groups
  /groups/{id}/activities:
    get:
      consumes:
      - application/json
      description: Get all activities for a group (members only)
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      - description: Requester User ID
        in: query
        name: requester_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/activity.Activity'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      summary: Get group activities
      tags:
      - groups
  /groups/{id}/invites:
    get:
      consumes:
      - application/json
      description: Get all invites for a group
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/group.GroupInvite'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      summary: Get group invites
      tags:
      - groups
    post:
      consumes:
      - application/json
      description: Create an invite link for the group (only group creator can create
        invites)
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      - description: Create invite request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/responses.CreateInviteRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/group.GroupInvite'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      summary: Create group invite link
      tags:
      - groups
  /groups/{id}/members:
    delete:
      consumes:
      - application/json
      description: Remove a user from a group (only group creator or the user themselves
        can remove)
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      - description: Remove user request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/responses.RemoveUserFromGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      summary: Remove user from group
      tags:
      - groups
    get:
      consumes:
      - application/json
      description: Get all members of a group (members only)
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      - description: Requester User ID
        in: query
        name: requester_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.GroupMembersResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      summary: Get group members
      tags:
      - groups
    post:
      consumes:
      - application/json
      description: Add a user to a group
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      - description: Add user request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/responses.AddUserToGroupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/responses.AddUserToGroupResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      summary: Add user to group
      tags:
      - groups
  /groups/{id}/members/nickname:
    delete:
      consumes:
      - application/json
      description: Remove a user's nickname in a group
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      - description: Delete nickname request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/responses.DeleteNicknameRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.SuccessResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      summary: Delete user nickname in group
      tags:
      - groups
    put:
      consumes:
      - application/json
      description: Set or update a user's nickname in a group
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      - description: Set nickname request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/responses.SetNicknameRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      summary: Set user nickname in group
      tags:
      - groups
  /invites/{invite_code}/deactivate:
    delete:
      consumes:
      - application/json
      description: Deactivate an invite link (only group creator can deactivate)
      parameters:
      - description: Invite Code
        in: path
        name: invite_code
        required: true
        type: string
      - description: Deactivate invite request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/responses.DeactivateInviteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.SuccessResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      summary: Deactivate group invite
      tags:
      - groups
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"